AG-20663 & AG-20662: fixed inputs for stealth cookies time & fixed colors in filtering log
Merge in EXTENSIONS/browser-extension from fixes/AG-20663 to master

Squashed commit of the following:

commit ed5657350991aa72b85fbfefa39279e201906aa3
Author: Dmitriy Seregin <d.seregin@adguard.com>
Date:   Tue Mar 28 00:01:39 2023 +0400

    fixed yarn.lock

commit 5c29712163f3fed8e2a728af0dd858453165bc86
Merge: dba5b7009 8a57c2aaa
Author: Dmitriy Seregin <d.seregin@adguard.com>
Date:   Mon Mar 27 23:58:18 2023 +0400

    Merge branch 'master' into fixes/AG-20663

commit dba5b70094f03627a1e8d5c6ac7d8ac7256e09d9
Author: Dmitriy Seregin <d.seregin@adguard.com>
Date:   Mon Mar 27 23:45:41 2023 +0400

    fixes

commit b6a5570cf41f6a7a64aef03105a68d1903030aef
Author: Dmitriy Seregin <d.seregin@adguard.com>
Date:   Mon Mar 27 17:08:22 2023 +0400

    AG-20663: fixed inputs for stealth cookies time

+++ Extension/src/background/schema/configuration/stealth.ts
@@ -77,7 +77,7 @@ export const stealthConfigValidator = zod.object({
     /**
      * Time in milliseconds to delete third-party cookies.
      */
-    [StealthOption.SelfDestructThirdPartyCookiesTime]: zod.number().int(),
+    [StealthOption.SelfDestructThirdPartyCookiesTime]: zod.number(),
     /**
      * Whether or not the application should set a fixed lifetime from
      * `StealthOption.SelfDestructFirstPartyCookiesTime` for first-party
@@ -87,7 +87,7 @@ export const stealthConfigValidator = zod.object({
     /**
      * Time in milliseconds to delete first-party cookies.
      */
-    [StealthOption.SelfDestructFirstPartyCookiesTime]: zod.number().int(),
+    [StealthOption.SelfDestructFirstPartyCookiesTime]: zod.number(),
     /**
      * If true application will enable AdGuard Tracking Protection filter
      * {@link AntiBannerFiltersId.TrackingFilterId}.
